+++ includes/Loop.php
@@ -2,7 +2,7 @@
 /**
  * WooCommerce loop (products lists).
  *
- * @package    Theme Integration for WooCommerce
+ * @package    Integration for WooCommerce
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0